requirements.txt
# Crypto Trading Bot Dependencies
sqlalchemy>=2.0.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
flask>=3.0.0
websocket-client>=1.6.0
rich>=13.7.1
